# indices: switch to worker/channel pattern