With GOMAXPROCS=2

Image: `moov/watchman:dev-d2af258`

<details>
<summary>Local benchmarks</summary>

```
                               │  before.txt  │              after.txt               │
                               │    sec/op    │    sec/op     vs base                │
API_Search/normal-2              22.24m ± ∞ ¹   22.80m ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-2               167.9m ± ∞ ¹   140.2m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-2     44.98m ± ∞ ¹   47.52m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/3-2     29.16m ± ∞ ¹   25.61m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/5-2     27.14m ± ∞ ¹   25.22m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/10-2    24.46m ± ∞ ¹   25.40m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/20-2    24.19m ± ∞ ¹   25.08m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/25-2    24.82m ± ∞ ¹   26.21m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/50-2    23.11m ± ∞ ¹   25.21m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-2   30.63m ± ∞ ¹   24.74m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/150-2   24.29m ± ∞ ¹   25.75m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-2   23.34m ± ∞ ¹   25.37m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/250-2   27.65m ± ∞ ¹   25.58m ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-2     24.51m ± ∞ ¹   26.26m ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/individual-2     62.55m ± ∞ ¹   61.23m ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/business-2       40.03m ± ∞ ¹   42.66m ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/vessel-2         15.91m ± ∞ ¹   16.59m ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/aircraft-2       16.55m ± ∞ ¹   16.29m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                          29.83m         29.67m        -0.56%

                               │  before.txt   │               after.txt               │
                               │     B/op      │     B/op       vs base                │
API_Search/normal-2              2.797Mi ± ∞ ¹   2.928Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-2               46.15Mi ± ∞ ¹   46.28Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-2     4.209Mi ± ∞ ¹   4.342Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/3-2     4.209Mi ± ∞ ¹   4.341Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/5-2     4.209Mi ± ∞ ¹   4.342Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/10-2    4.210Mi ± ∞ ¹   4.342Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/20-2    4.211Mi ± ∞ ¹   4.342Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/25-2    4.211Mi ± ∞ ¹   4.343Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/50-2    4.214Mi ± ∞ ¹   4.345Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-2   4.218Mi ± ∞ ¹   4.348Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/150-2   4.223Mi ± ∞ ¹   4.351Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-2   4.227Mi ± ∞ ¹   4.354Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/250-2   4.232Mi ± ∞ ¹   4.357Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-2     4.211Mi ± ∞ ¹   4.342Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/individual-2     12.12Mi ± ∞ ¹   12.25Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/business-2       9.003Mi ± ∞ ¹   9.135Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/vessel-2         2.169Mi ± ∞ ¹   2.301Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/aircraft-2       2.102Mi ± ∞ ¹   2.234Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                          4.827Mi         4.979Mi        +3.16%

                               │  before.txt  │              after.txt               │
                               │  allocs/op   │  allocs/op    vs base                │
API_Search/normal-2              189.1k ± ∞ ¹   189.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
API_Search/debug-2               411.0k ± ∞ ¹   411.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/1-2     280.0k ± ∞ ¹   280.0k ± ∞ ¹       ~ (p=1.000 n=1) ³
_Search/fixed_group_size/3-2     280.0k ± ∞ ¹   280.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/5-2     280.0k ± ∞ ¹   280.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/10-2    280.1k ± ∞ ¹   280.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/20-2    280.1k ± ∞ ¹   280.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/25-2    280.1k ± ∞ ¹   280.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/50-2    280.1k ± ∞ ¹   280.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/100-2   280.2k ± ∞ ¹   280.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/150-2   280.3k ± ∞ ¹   280.2k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/200-2   280.4k ± ∞ ¹   280.2k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/fixed_group_size/250-2   280.5k ± ∞ ¹   280.3k ± ∞ ¹       ~ (p=1.000 n=1) ²
_Search/dynamic_group_size-2     280.1k ± ∞ ¹   280.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/individual-2     811.9k ± ∞ ¹   811.9k ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/business-2       456.9k ± ∞ ¹   456.9k ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/vessel-2         183.1k ± ∞ ¹   183.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
_SearchParallel/aircraft-2       183.1k ± ∞ ¹   183.1k ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                          291.2k         291.1k        -0.02%
```

</details>